Refactor code structure for improved readability and maintainability Finished function List Vocab, can add custom vocab and dictionary vocab into list

+++ src/main/java/com/englishvocab/entity/Media.java
@@ -25,7 +25,7 @@ public class Media {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long mediaId;
+    Integer mediaId;  // Changed from Long to Integer for consistency with other entities
     
     @Enumerated(EnumType.STRING)
     @Column(nullable = false, length = 50)
